+++ nova_web/src/lib.rs
@@ -7,3 +7,9 @@
 
 /// nova server implementation
 pub mod server;
+
+/// nova core
+pub mod core;
+
+/// nova routing
+pub mod routing;
