+++ nova_web/src/server.rs
@@ -7,19 +7,15 @@ use nova_core::errors::ServerError;
 use nova_core::request::HttpRequest;
 use nova_core::response::HttpResponse;
 use nova_core::types::protocol::Protocol;
-use nova_core::types::request_type::RequestType;
 
-use nova_router::callable::{CloneableFn, ServerResponse};
-use nova_router::route::Route;
 use nova_router::router::Router;
-use nova_router::server_routing::ServerRouting;
 
 /// Nova server structure
 #[derive(Clone, Debug)]
 pub struct Server {
     host: String,
     port: u16,
-    router: Router,
+    pub(crate) router: Router,
     protocol: Protocol,
 }
 
@@ -83,29 +79,5 @@ impl Server {
         let response = HttpResponse::from_error(error, Protocol::default())
             .append_default_headers();
         stream.write_all(format!("{response}").as_bytes()).await
-    }
-}
-
-impl ServerRouting for Server {
-    fn route<F: CloneableFn<Output=ServerResponse> + 'static>(&mut self, r#type: RequestType, path: &str, f: F) -> Self {
-        self.router.register(r#type, path, f);
-        self.clone()
-    }
-
-    fn service(&mut self, path: &str, routes: Vec<Route>) -> Self where Self: Sized {
-        routes.into_iter().for_each(|item| {
-            let path = &format!("{path}{}", item.clone().get_path());
-            if item.clone().get_callable().is_some() {
-                self.route(item.clone().get_type(), path, item.get_callable().unwrap());
-            } else {
-                self.service(path, item.get_routes());
-            }
-        });
-        self.clone()
-    }
-
-    fn fallback<F: CloneableFn<Output=ServerResponse> + 'static>(&mut self, f: F) -> Self where Self: Sized {
-        self.router.register_fallback(f);
-        self.clone()
     }
 }
